requirements.txt
docker>=7.1.0
ttkbootstrap>=1.20.1
pillow>=12.1.0
maplex>=3.0.0b1